src/main/java/org/igye/outline/typeconverters/ListConverter.java
package org.igye.outline.typeconverters;

import org.apache.commons.lang3.StringUtils;

import javax.persistence.AttributeConverter;
import java.util.Arrays;
import java.util.Collections;
import java.util.List;

public class ListConverter implements AttributeConverter<List<String>, String> {
    @Override
    public String convertToDatabaseColumn(List<String> attribute) {
        return StringUtils.join(attribute, ";");
    }

    @Override
    public List<String> convertToEntityAttribute(String dbData) {
        if (StringUtils.isEmpty(dbData)) {
            return Collections.emptyList();
        } else {
            return Arrays.asList(dbData.split(";"));
        }
    }
}
